requirement.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
psycopg2-binary
pyodbc
